+++ CHANGELOG.md
@@ -5,6 +5,7 @@ tag k0.11.5
 - The subtitle delay API is incompatible, the value is need reverse: the positive value means subtitle put off, the negative value means subtitle advanced
 - optimized the continuous seek logic
 - fix crash: target thread exited while waiting for the perform
+- support specify fonts dir for subtitle
 
 tag k0.11.4
 --------------------------------